requirements.txt
Flask-Login
Flask-SQLAlchemy
OWSLib
